Refactored Inventory Activity Files
