+++ app/src/main/java/com/viiishoppinglistapp/doit/ui/inventory/PageViewModelInventory.java
@@ -1,4 +1,4 @@
-package com.viiishoppinglistapp.doit.ui.main;
+package com.viiishoppinglistapp.doit.ui.inventory;
 
 import androidx.arch.core.util.Function;
 import androidx.lifecycle.LiveData;
@@ -6,7 +6,7 @@ import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.Transformations;
 import androidx.lifecycle.ViewModel;
 
-public class PageViewModel extends ViewModel {
+public class PageViewModelInventory extends ViewModel {
 
     private MutableLiveData<Integer> mIndex = new MutableLiveData<>();
     private LiveData<String> mText = Transformations.map(mIndex, new Function<Integer, String>() {
